Fold run_common into tahoe_run since there are no other run-like commands anymore
